requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic>=2.0.0
openai>=1.0.0
python-dotenv==1.0.0
pydantic>=2.0.0
flask>=2.2.3
werkzeug>=2.2.3
pdfkit>=1.0.0
uuid>=1.30
